Removed unnecessary explicit conversion

+++ 06 - Sub-rotinas/serie de numeros.c
@@ -5,7 +5,7 @@ double calculaSerie(int n) {
     double s = 0;
 
     for (i = 1; i <= n; i += 2)
-        s += (double)(topo / pow(i, 2)) * (topo / pow(i + 1, 2)), topo++;
+        s += (topo / pow(i, 2)) * (topo / pow(i + 1, 2)), topo++;
 
     return s;
 }
